fix: ublox m9 file names
