+++ drivers/gnss/ublox_neo_m9n/CMakeLists.txt
@@ -1,0 +1,3 @@
+zephyr_library()
+zephyr_library_sources_ifdef(CONFIG_UBLOX_NEO_M9N src/ublox_neo_m9n.c)
+zephyr_library_include_directories(include)
