+++ drivers/gnss/ublox_neo_m9n/dts/bindings/gnss/ublox,neom9n.yaml
